# openapi.yaml
openapi: 3.0.3
info:
  title: API de Charadas
  description: API para gerenciamento de charadas com integração Firebase e autenticação via Token.
  version: 1.0.0
  contact:
    name: Barros
servers:
  - url: https://api-charadas-eosin.vercel.app/
    description: Servidor de Produção (Vercel)
  - url: http://127.0.0.1:5000
    description: Servidor Local

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Insira o token obtido na rota /login para acessar rotas protegidas."

  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pergunta:
          type: string
          example: "O que é, o que é?"
        resposta:
          type: string
          example: "É o que é."
    
    CharadaInput:
      type: object
      required:
        - pergunta
        - resposta
      properties:
        pergunta:
          type: string
        resposta:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string

paths:
  /:
    get:
      summary: Informações da API
      responses:
        '200':
          description: Retorna metadados da API.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api: { type: string }
                  version: { type: string }
                  author: { type: string }

  /login:
    post:
      summary: Realiza login para obter token de acesso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usuario
                - senha
              properties:
                usuario: { type: string }
                senha: { type: string }
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  token: { type: string }
        '400':
          description: Dados insuficientes
        '401':
          description: Credenciais inválidas

  /charadas:
    get:
      summary: Lista todas as charadas
      responses:
        '200':
          description: Lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'
    
    post:
      summary: Adiciona uma nova charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaInput'
      responses:
        '201':
          description: Criada com sucesso
        '400':
          description: Dados inválidos
        '500':
          description: Erro interno

  /charadas/random:
    get:
      summary: Retorna uma charada aleatória
      responses:
        '200':
          description: Uma charada selecionada aleatoriamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Nenhuma charada encontrada

  /charadas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    
    get:
      summary: Busca uma charada por ID
      responses:
        '200':
          description: Charada encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Charada não encontrada

    put:
      summary: Atualiza uma charada (Completo)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaInput'
      responses:
        '200':
          description: Atualizada com sucesso
        '404':
          description: Charada não encontrada

    patch:
      summary: Atualiza parcialmente uma charada
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta: { type: string }
                resposta: { type: string }
      responses:
        '200':
          description: Atualizada com sucesso

    delete:
      summary: Remove uma charada
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Apagada com sucesso
        '404':
          description: Charada não encontrada